refactor(rag): modularize RAG service into focused submodules
- Extract config/constants to rag_config.py
- Extract embedding management to rag_embeddings.py
- Extract ingest/document processing to rag_ingest.py
- Extract retrieval pipeline to rag_retrieval.py
- Extract summarization to rag_summarization.py
- Extract hybrid search (HyDE, BM25, RRF, PDR) to rag_hybrid.py
- Extract web search policy to rag_policy.py
- Keep rag_service.py as facade for backward compatibility

Total: 1956 → 1753 lines across 8 modules
